requirements.txt
streamlit==1.31.1
python-dotenv==1.0.0
groq==0.4.1
duckduckgo-search==4.1.1
requests==2.31.0
urllib3==2.0.7
pydantic==2.5.3
typing-extensions==4.8.0